PIKURR/requirements.txt
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
pandas
numpy
rasterio
shapely
scikit-image
tqdm
requests
tenacity
google-auth
earthengine-api
ovmsclient
Pillow
python-dotenv
streamlit